Added pagination to the domains table.
