--- app/Http/Livewire/DomainTableEmpty.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Livewire;
-
-use Livewire\Component;
-
-class DomainTableEmpty extends Component
-{
-    public function render()
-    {
-        return view('livewire.domain-table-empty');
-    }
-}
