--- resources/views/livewire/domain-table-empty.blade.php
@@ -1,7 +1,0 @@
-<tr>
-    <td colspan="8" class="text-sm font-medium whitespace-nowrap text-center px-6 py-4 text-gray-700">
-        <span>
-            There doesn't appear to be any domains, why not <a class="underline" href="{{ route('domains.create') }}">Add one?</a>
-        </span>
-    </td>
-</tr>
