+++ resources/views/livewire/search-bar.blade.php
@@ -1,7 +1,9 @@
-<div class="relative text-gray-600">
-    <input class="border-1 shadow-sm border-gray-200 bg-white px-5 pr-16 rounded-lg text-sm focus:outline-none"
-           type="search" name="search" placeholder="Search">
-    <button type="submit" class="absolute right-0 top-0 mt-2.5 mr-4">
-        <x-heroicon-o-search class="text-gray-600 h-5 w-5"/>
-    </button>
+<div>
+    <div class="relative text-gray-600">
+        <input class="border-1 shadow-sm border-gray-200 bg-white px-5 pr-16 rounded-lg text-sm focus:outline-none"
+               type="search" name="search" placeholder="Search">
+        <button type="submit" class="absolute right-0 top-0 mt-2.5 mr-4">
+            <x-heroicon-o-search class="text-gray-600 h-5 w-5"/>
+        </button>
+    </div>
 </div>
